bad commit

+++ loader.go
@@ -5,7 +5,7 @@ type Loader interface {
 	Start()
 	// TODO: Build on this and make the two types similar enough that we can use
 	// this loader interface instead of typing out to their type
-	Increment(int) bool
+	Increment(float64) bool
 	End()
 }
